+++ crates/voicevox_core_c_api/tests/e2e/log_mask.rs
@@ -23,7 +23,8 @@ impl Utf8Output {
     }
 
     pub(crate) fn mask_onnxruntime_filename(self) -> Self {
-        const ONNXRUNTIME_VERSION: &str = include_str!("../../../../onnxruntime-version.txt");
+        const ONNXRUNTIME_VERSION: &str =
+            include_str!("../../../voicevox_core_build_features/onnxruntime-version-for-test.txt");
         self.mask_stderr(
             static_regex!(regex::escape(
                 const {